move dd_mock and mock_customer to conftest
