+++ featuretools/tests/conftest.py
@@ -155,3 +155,16 @@ def games_es(home_games_es):
 @pytest.fixture
 def pd_mock_customer():
     return ft.demo.load_mock_customer(return_entityset=True, random_seed=0)
+
+
+@pytest.fixture
+def dd_mock_customer(pd_mock_customer):
+    dd_mock_customer = copy.deepcopy(pd_mock_customer)
+    for entity in dd_mock_customer.entities:
+        entity.df = dd.from_pandas(entity.df.reset_index(drop=True), npartitions=4)
+    return dd_mock_customer
+
+
+@pytest.fixture(params=['pd_mock_customer', 'dd_mock_customer'])
+def mock_customer(request):
+    return request.getfixturevalue(request.param)
